requirements.txt
streamlit
inference-sdk
pillow
python-dotenv
git+https://github.com/openai/CLIP.git
torch
torchvision
huggingface-hub
transformers